Implemented more on outbound transactions and their timeouts.
The ActiveOutboundTransactions <-> OutboundTransactions have grown a lot, and it's rather complex now... Even though it's a lot
simpler than it was a few commits ago (if I commited that at all...). Anyway: Most things are now handled in the thread where the
task is created, and only a simple mapping between SID and task (which implies a thread) is really global and is protected with
a mutex. The rest is single-threaded and simple.

Weak pointers are used in all places to ensure that our reference counting really works. And it has to work, because
we are depending that transactions clean up after themselves when they are deleted. If not, assertions will let us know.