Removed timer.h/cpp because we have changed how they work.
The old timer solution was a flexible and generic solution, that complicated things.
When timers only are used for timing out transactions, and they are task/thread private
and not considered events, this solution wasn't useful.

So it will be two parts: Transaction timeouts - handled internally by the Transaction module,
and task scheduling at certain intervals (or delays). To be implemented.